lab1/second_part/requirements.txt
hopsworks
joblib
scikit-learn
seaborn
pandas
numpy
dataframe-image
modal-client